Edit VERSION in config file (#257)
